+++ src/core.ts
@@ -2,6 +2,7 @@ import {
   serialize as defaultSerializer,
   deserialize as defaultDeserializer,
   addGlobalAllowedClass,
+  type ClassDefinition,
 } from "@macfja/serializer"
 import { get as getCookie, set as setCookie, erase as removeCookie } from "browser-cookies"
 import { get, set, createStore, del } from "idb-keyval"
@@ -48,7 +49,7 @@ const warnStorageNotFound = (storageName) => {
  * Add a class to the allowed list of classes to be serialized
  * @param classDef The class to add to the list
  */
-export const addSerializableClass = (classDef: FunctionConstructor): void => {
+export const addSerializableClass = (classDef: ClassDefinition<any>): void => {
   addSerializable(classDef)
 }
 
@@ -68,7 +69,7 @@ export let serialize = defaultSerializer
 export let deserialize = defaultDeserializer
 /**
  * The function used to add a class in the serializer allowed class
- * @type {(classConstructor: FunctionConstructor) => void}
+ * @type {(classConstructor: ClassDefinition<any>) => void}
  */
 let addSerializable = addGlobalAllowedClass
 
@@ -76,12 +77,12 @@ let addSerializable = addGlobalAllowedClass
  * Set the serialization functions to use
  * @param {(data: any) => string} serializer The function to use to transform any data into a string
  * @param {(input: string) => any} deserializer The function to use to transform back string into the original data (reverse of the serializer)
- * @param {(classConstructor: FunctionConstructor) => void} [addSerializableClass] The function to use to add a class in the serializer/deserializer allowed class
+ * @param {(classConstructor: ClassDefinition<any>) => void} [addSerializableClass] The function to use to add a class in the serializer/deserializer allowed class
  */
 export function setSerialization(
   serializer: (data: any) => string,
   deserializer: (input: string) => any,
-  addSerializableClass?: (classConstructor: FunctionConstructor) => void
+  addSerializableClass?: (classConstructor: ClassDefinition<any>) => void
 ): void {
   serialize = serializer
   deserialize = deserializer